SatriaFinaceelTest/requirements.txt
urllib3==1.25.9
robotframework==3.1.1
robotframework-appiumlibrary==1.4.6
robotframework-debuglibrary==1.1.4
robotframework-faker==4.2.0
robotframework-pabot==0.91
robotframework-requests==0.5.0
robotframework-seleniumlibrary==4.1.0
RESTinstance==1.0.2
selenium==3.141.0